refc: More structure reformatting
